can now select channel in instrument editor can also set envelope/waveform sequences can now be edited via line edit EnvelopeForm now shows the rate of decay/attack for the set period ModuleModel forwards dataChanged signal from child models Readded SequenceModel class Fixed sign issue in FrequenceControl for negative pitch offsets

+++ libtrackerboy/engine/FrequencyControl.cpp
@@ -255,7 +255,7 @@ void FrequencyControl::step() noexcept {
     if (mContext) {
         auto pitch = mContext->pitchSequence.next();
         if (pitch) {
-            mInstrumentPitch += *pitch;
+            mInstrumentPitch += (int8_t)*pitch;
         }
 
         arp = mContext->arpSequence.next();
